Replace GaussianGRUPolicy (#922)
Replace GaussianGRUPolicy with model based implementation and add
corresponding tests and benchmark script.
