+++ tests/garage/tf/algos/test_ppo_with_model.py
@@ -10,7 +10,7 @@ from garage.tf.algos import PPO
 from garage.tf.baselines import GaussianMLPBaseline
 from garage.tf.envs import TfEnv
 from garage.tf.experiment import LocalTFRunner
-from garage.tf.policies import GaussianGRUPolicyWithModel
+from garage.tf.policies import GaussianGRUPolicy
 from tests.fixtures import snapshot_config, TfGraphTestCase
 
 
@@ -21,7 +21,7 @@ class TestPPOWithModel(TfGraphTestCase):
         """Test PPO with Pendulum environment and GRU policy."""
         with LocalTFRunner(snapshot_config, sess=self.sess) as runner:
             env = TfEnv(normalize(gym.make('InvertedDoublePendulum-v2')))
-            policy = GaussianGRUPolicyWithModel(env_spec=env.spec, )
+            policy = GaussianGRUPolicy(env_spec=env.spec, )
             baseline = GaussianMLPBaseline(
                 env_spec=env.spec,
                 regressor_args=dict(hidden_sizes=(32, 32)),
