+++ tests/garage/tf/models/test_gaussian_gru_model.py
@@ -11,6 +11,7 @@ from tests.helpers import recurrent_step_gru
 
 
 class TestGaussianGRUModel(TfGraphTestCase):
+
     def setup_method(self):
         super().setup_method()
         self.batch_size = 1
@@ -22,8 +23,10 @@ class TestGaussianGRUModel(TfGraphTestCase):
             (self.batch_size, self.time_step, self.feature_shape), 1.)
         self.obs_input = np.full((self.batch_size, self.feature_shape), 1.)
 
-        self.input_var = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, self.feature_shape), name='input')
+        self.input_var = tf.compat.v1.placeholder(tf.float32,
+                                                  shape=(None, None,
+                                                         self.feature_shape),
+                                                  name='input')
         self.step_input_var = tf.compat.v1.placeholder(
             tf.float32, shape=(None, self.feature_shape), name='step_input')
 
@@ -38,22 +41,22 @@ class TestGaussianGRUModel(TfGraphTestCase):
     def test_std_share_network_output_values(self, mock_normal, output_dim,
                                              hidden_dim):
         mock_normal.return_value = 0.5
-        model = GaussianGRUModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=True,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, hidden_init_var, dist) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var)
+        model = GaussianGRUModel(output_dim=output_dim,
+                                 hidden_dim=hidden_dim,
+                                 std_share_network=True,
+                                 hidden_nonlinearity=None,
+                                 recurrent_nonlinearity=None,
+                                 hidden_w_init=self.default_initializer,
+                                 recurrent_w_init=self.default_initializer,
+                                 output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         hidden_init_var, dist) = model.build(self.input_var,
+                                              self.step_input_var,
+                                              step_hidden_var)
 
         hidden1 = hidden2 = np.full((self.batch_size, hidden_dim),
                                     hidden_init_var.eval())
@@ -63,22 +66,21 @@ class TestGaussianGRUModel(TfGraphTestCase):
             feed_dict={self.input_var: self.obs_inputs})
 
         for i in range(self.time_step):
-            action, mean1, log_std1, hidden1 = self.sess.run(
-                [action_var, step_mean_var, step_log_std_var, step_hidden],
+            mean1, log_std1, hidden1 = self.sess.run(
+                [step_mean_var, step_log_std_var, step_hidden],
                 feed_dict={
                     self.step_input_var: self.obs_input,
                     step_hidden_var: hidden1
                 })
 
-            hidden2 = recurrent_step_gru(
-                input_val=self.obs_input,
-                num_units=hidden_dim,
-                step_hidden=hidden2,
-                w_x_init=0.1,
-                w_h_init=0.1,
-                b_init=0.,
-                nonlinearity=None,
-                gate_nonlinearity=None)
+            hidden2 = recurrent_step_gru(input_val=self.obs_input,
+                                         num_units=hidden_dim,
+                                         step_hidden=hidden2,
+                                         w_x_init=0.1,
+                                         w_h_init=0.1,
+                                         b_init=0.,
+                                         nonlinearity=None,
+                                         gate_nonlinearity=None)
 
             output_nonlinearity = np.full(
                 (np.prod(hidden2.shape[1:]), output_dim), 0.1)
@@ -86,9 +88,6 @@ class TestGaussianGRUModel(TfGraphTestCase):
             assert np.allclose(mean1, output2)
             assert np.allclose(log_std1, output2)
             assert np.allclose(hidden1, hidden2)
-
-            expected_action = 0.5 * np.exp(log_std1) + mean1
-            assert np.allclose(action, expected_action)
 
     # yapf: disable
     @pytest.mark.parametrize('output_dim, hidden_dim', [
@@ -98,22 +97,22 @@ class TestGaussianGRUModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_std_share_network_shapes(self, output_dim, hidden_dim):
-        model = GaussianGRUModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=True,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, hidden_init_var, dist) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var)
+        model = GaussianGRUModel(output_dim=output_dim,
+                                 hidden_dim=hidden_dim,
+                                 std_share_network=True,
+                                 hidden_nonlinearity=None,
+                                 recurrent_nonlinearity=None,
+                                 hidden_w_init=self.default_initializer,
+                                 recurrent_w_init=self.default_initializer,
+                                 output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         hidden_init_var, dist) = model.build(self.input_var,
+                                              self.step_input_var,
+                                              step_hidden_var)
 
         # output layer is a tf.keras.layers.Dense object,
         # which cannot be access by tf.compat.v1.variable_scope.
@@ -137,26 +136,27 @@ class TestGaussianGRUModel(TfGraphTestCase):
     ])
     # yapf: enable
     @mock.patch('tensorflow.random.normal')
-    def test_without_std_share_network_output_values(
-            self, mock_normal, output_dim, hidden_dim, init_std):
+    def test_without_std_share_network_output_values(self, mock_normal,
+                                                     output_dim, hidden_dim,
+                                                     init_std):
         mock_normal.return_value = 0.5
-        model = GaussianGRUModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=False,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer,
-            init_std=init_std)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, hidden_init_var, dist) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var)
+        model = GaussianGRUModel(output_dim=output_dim,
+                                 hidden_dim=hidden_dim,
+                                 std_share_network=False,
+                                 hidden_nonlinearity=None,
+                                 recurrent_nonlinearity=None,
+                                 hidden_w_init=self.default_initializer,
+                                 recurrent_w_init=self.default_initializer,
+                                 output_w_init=self.default_initializer,
+                                 init_std=init_std)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         hidden_init_var, dist) = model.build(self.input_var,
+                                              self.step_input_var,
+                                              step_hidden_var)
 
         hidden1 = hidden2 = np.full((self.batch_size, hidden_dim),
                                     hidden_init_var.eval())
@@ -166,22 +166,21 @@ class TestGaussianGRUModel(TfGraphTestCase):
             feed_dict={self.input_var: self.obs_inputs})
 
         for i in range(self.time_step):
-            action, mean1, log_std1, hidden1 = self.sess.run(
-                [action_var, step_mean_var, step_log_std_var, step_hidden],
+            mean1, log_std1, hidden1 = self.sess.run(
+                [step_mean_var, step_log_std_var, step_hidden],
                 feed_dict={
                     self.step_input_var: self.obs_input,
                     step_hidden_var: hidden1
                 })
 
-            hidden2 = recurrent_step_gru(
-                input_val=self.obs_input,
-                num_units=hidden_dim,
-                step_hidden=hidden2,
-                w_x_init=0.1,
-                w_h_init=0.1,
-                b_init=0.,
-                nonlinearity=None,
-                gate_nonlinearity=None)
+            hidden2 = recurrent_step_gru(input_val=self.obs_input,
+                                         num_units=hidden_dim,
+                                         step_hidden=hidden2,
+                                         w_x_init=0.1,
+                                         w_h_init=0.1,
+                                         b_init=0.,
+                                         nonlinearity=None,
+                                         gate_nonlinearity=None)
 
             output_nonlinearity = np.full(
                 (np.prod(hidden2.shape[1:]), output_dim), 0.1)
@@ -192,9 +191,6 @@ class TestGaussianGRUModel(TfGraphTestCase):
             assert np.allclose(log_std1, expected_log_std)
             assert np.allclose(hidden1, hidden2)
 
-            expected_action = 0.5 * np.exp(log_std1) + mean1
-            assert np.allclose(action, expected_action)
-
     # yapf: disable
     @pytest.mark.parametrize('output_dim, hidden_dim', [
         (1, 1),
@@ -203,22 +199,22 @@ class TestGaussianGRUModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_without_std_share_network_shapes(self, output_dim, hidden_dim):
-        model = GaussianGRUModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=False,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, hidden_init_var, dist) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var)
+        model = GaussianGRUModel(output_dim=output_dim,
+                                 hidden_dim=hidden_dim,
+                                 std_share_network=False,
+                                 hidden_nonlinearity=None,
+                                 recurrent_nonlinearity=None,
+                                 hidden_w_init=self.default_initializer,
+                                 recurrent_w_init=self.default_initializer,
+                                 output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         hidden_init_var, dist) = model.build(self.input_var,
+                                              self.step_input_var,
+                                              step_hidden_var)
 
         # output layer is a tf.keras.layers.Dense object,
         # which cannot be access by tf.compat.v1.variable_scope.
@@ -245,22 +241,21 @@ class TestGaussianGRUModel(TfGraphTestCase):
     def test_std_share_network_is_pickleable(self, mock_normal, output_dim,
                                              hidden_dim):
         mock_normal.return_value = 0.5
-        model = GaussianGRUModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=True,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        (_, mean_var, step_mean_var, log_std_var,
-         step_log_std_var, step_hidden, _, _) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var)
+        model = GaussianGRUModel(output_dim=output_dim,
+                                 hidden_dim=hidden_dim,
+                                 std_share_network=True,
+                                 hidden_nonlinearity=None,
+                                 recurrent_nonlinearity=None,
+                                 hidden_w_init=self.default_initializer,
+                                 recurrent_w_init=self.default_initializer,
+                                 output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         _, _) = model.build(self.input_var, self.step_input_var,
+                             step_hidden_var)
 
         # output layer is a tf.keras.layers.Dense object,
         # which cannot be access by tf.compat.v1.variable_scope.
@@ -283,22 +278,23 @@ class TestGaussianGRUModel(TfGraphTestCase):
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             model_pickled = pickle.loads(h)
 
-            input_var = tf.compat.v1.placeholder(
-                tf.float32,
-                shape=(None, None, self.feature_shape),
-                name='input')
+            input_var = tf.compat.v1.placeholder(tf.float32,
+                                                 shape=(None, None,
+                                                        self.feature_shape),
+                                                 name='input')
             step_input_var = tf.compat.v1.placeholder(
                 tf.float32,
                 shape=(None, self.feature_shape),
                 name='step_input')
-            step_hidden_var = tf.compat.v1.placeholder(
-                shape=(self.batch_size, hidden_dim),
-                name='initial_hidden',
-                dtype=tf.float32)
+            step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                              hidden_dim),
+                                                       name='initial_hidden',
+                                                       dtype=tf.float32)
 
-            (_, mean_var2, step_mean_var2, log_std_var2,
-             step_log_std_var2, step_hidden2, _, _) = model_pickled.build(
-                 input_var, step_input_var, step_hidden_var)
+            (mean_var2, step_mean_var2, log_std_var2, step_log_std_var2,
+             step_hidden2, _, _) = model_pickled.build(input_var,
+                                                       step_input_var,
+                                                       step_hidden_var)
 
             outputs2 = sess.run([mean_var2, log_std_var2],
                                 feed_dict={input_var: self.obs_inputs})
@@ -322,22 +318,21 @@ class TestGaussianGRUModel(TfGraphTestCase):
     def test_without_std_share_network_is_pickleable(self, mock_normal,
                                                      output_dim, hidden_dim):
         mock_normal.return_value = 0.5
-        model = GaussianGRUModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=False,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        (_, mean_var, step_mean_var, log_std_var,
-         step_log_std_var, step_hidden, _, _) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var)
+        model = GaussianGRUModel(output_dim=output_dim,
+                                 hidden_dim=hidden_dim,
+                                 std_share_network=False,
+                                 hidden_nonlinearity=None,
+                                 recurrent_nonlinearity=None,
+                                 hidden_w_init=self.default_initializer,
+                                 recurrent_w_init=self.default_initializer,
+                                 output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         _, _) = model.build(self.input_var, self.step_input_var,
+                             step_hidden_var)
 
         # output layer is a tf.keras.layers.Dense object,
         # which cannot be access by tf.compat.v1.variable_scope.
@@ -360,22 +355,23 @@ class TestGaussianGRUModel(TfGraphTestCase):
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             model_pickled = pickle.loads(h)
 
-            input_var = tf.compat.v1.placeholder(
-                tf.float32,
-                shape=(None, None, self.feature_shape),
-                name='input')
+            input_var = tf.compat.v1.placeholder(tf.float32,
+                                                 shape=(None, None,
+                                                        self.feature_shape),
+                                                 name='input')
             step_input_var = tf.compat.v1.placeholder(
                 tf.float32,
                 shape=(None, self.feature_shape),
                 name='step_input')
-            step_hidden_var = tf.compat.v1.placeholder(
-                shape=(self.batch_size, hidden_dim),
-                name='initial_hidden',
-                dtype=tf.float32)
+            step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                              hidden_dim),
+                                                       name='initial_hidden',
+                                                       dtype=tf.float32)
 
-            (_, mean_var2, step_mean_var2, log_std_var2,
-             step_log_std_var2, step_hidden2, _, _) = model_pickled.build(
-                 input_var, step_input_var, step_hidden_var)
+            (mean_var2, step_mean_var2, log_std_var2, step_log_std_var2,
+             step_hidden2, _, _) = model_pickled.build(input_var,
+                                                       step_input_var,
+                                                       step_hidden_var)
 
             outputs2 = sess.run([mean_var2, log_std_var2],
                                 feed_dict={input_var: self.obs_inputs})
