+++ tests/garage/tf/policies/test_gaussian_gru_policy.py
@@ -6,14 +6,15 @@ import pytest
 import tensorflow as tf
 
 from garage.tf.envs import TfEnv
-from garage.tf.policies import GaussianGRUPolicyWithModel
+from garage.tf.policies import GaussianGRUPolicy
 from tests.fixtures import TfGraphTestCase
 from tests.fixtures.envs.dummy import DummyBoxEnv
 from tests.fixtures.envs.dummy import DummyDiscreteEnv
 from tests.fixtures.models import SimpleGaussianGRUModel
 
 
-class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
+class TestGaussianGRUPolicy(TfGraphTestCase):
+
     @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
         ((1, ), (1, ), 4),
         ((2, ), (2, ), 4),
@@ -27,16 +28,17 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
             tf.float32, shape=(None, None, env.observation_space.flat_dim))
 
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy_with_model.GaussianGRUModel'),
+                         'gaussian_gru_policy.GaussianGRUModel'),
                         new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicyWithModel(
-                env_spec=env.spec, state_include_action=False)
+            policy = GaussianGRUPolicy(env_spec=env.spec,
+                                       state_include_action=False)
 
             policy.reset()
             obs = env.reset()
 
-            dist_sym = policy.dist_info_sym(
-                obs_var=obs_ph, state_info_vars=None, name='p2_sym')
+            dist_sym = policy.dist_info_sym(obs_var=obs_ph,
+                                            state_info_vars=None,
+                                            name='p2_sym')
 
         dist = self.sess.run(
             dist_sym, feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
@@ -59,10 +61,10 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
             tf.float32, shape=(None, None, env.observation_space.flat_dim))
 
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy_with_model.GaussianGRUModel'),
+                         'gaussian_gru_policy.GaussianGRUModel'),
                         new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicyWithModel(
-                env_spec=env.spec, state_include_action=True)
+            policy = GaussianGRUPolicy(env_spec=env.spec,
+                                       state_include_action=True)
 
             policy.reset()
             obs = env.reset()
@@ -84,10 +86,10 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
             tf.float32, shape=(None, None, env.observation_space.flat_dim))
 
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy_with_model.GaussianGRUModel'),
+                         'gaussian_gru_policy.GaussianGRUModel'),
                         new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicyWithModel(
-                env_spec=env.spec, state_include_action=True)
+            policy = GaussianGRUPolicy(env_spec=env.spec,
+                                       state_include_action=True)
 
             policy.reset()
             obs = env.reset()
@@ -105,32 +107,35 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
     def test_invalid_env(self):
         env = TfEnv(DummyDiscreteEnv())
         with pytest.raises(ValueError):
-            GaussianGRUPolicyWithModel(env_spec=env.spec)
+            GaussianGRUPolicy(env_spec=env.spec)
 
     # yapf: disable
+
     @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
         ((1, ), (1, ), 4),
         ((2, ), (2, ), 4),
         ((1, 1), (1, ), 4),
         ((2, 2), (2, ), 4)
     ])
+    @mock.patch('numpy.random.normal')
     # yapf: enable
-    def test_get_action_state_include_action(self, obs_dim, action_dim,
-                                             hidden_dim):
-
+    def test_get_action_state_include_action(self, mock_normal, obs_dim,
+                                             action_dim, hidden_dim):
+        mock_normal.return_value = 0.5
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy_with_model.GaussianGRUModel'),
+                         'gaussian_gru_policy.GaussianGRUModel'),
                         new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicyWithModel(
-                env_spec=env.spec, state_include_action=True)
+            policy = GaussianGRUPolicy(env_spec=env.spec,
+                                       state_include_action=True)
 
         policy.reset()
         obs = env.reset()
 
+        expected_action = np.full(action_dim, 0.5 * np.exp(0.5) + 0.5)
         action, agent_info = policy.get_action(obs)
         assert env.action_space.contains(action)
-        assert np.array_equal(action, np.full(action_dim, 0.75))
+        assert np.allclose(action, expected_action)
         expected_mean = np.full(action_dim, 0.5)
         assert np.array_equal(agent_info['mean'], expected_mean)
         expected_log_std = np.full(action_dim, 0.5)
@@ -145,7 +150,7 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
                 actions, agent_infos['mean'], agent_infos['log_std'],
                 agent_infos['prev_action']):
             assert env.action_space.contains(action)
-            assert np.array_equal(action, np.full(action_dim, 0.75))
+            assert np.allclose(action, expected_action)
             assert np.array_equal(mean, expected_mean)
             assert np.array_equal(log_std, expected_log_std)
             assert np.array_equal(prev_action, expected_prev_action)
@@ -157,21 +162,25 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
         ((1, 1), (1, ), 4),
         ((2, 2), (2, ), 4)
     ])
+    @mock.patch('numpy.random.normal')
     # yapf: enable
-    def test_get_action(self, obs_dim, action_dim, hidden_dim):
+    def test_get_action(self, mock_normal, obs_dim, action_dim, hidden_dim):
+        mock_normal.return_value = 0.5
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy_with_model.GaussianGRUModel'),
+                         'gaussian_gru_policy.GaussianGRUModel'),
                         new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicyWithModel(
-                env_spec=env.spec, state_include_action=False)
+            policy = GaussianGRUPolicy(env_spec=env.spec,
+                                       state_include_action=False)
 
         policy.reset()
         obs = env.reset()
 
+        expected_action = np.full(action_dim, 0.5 * np.exp(0.5) + 0.5)
         action, agent_info = policy.get_action(obs)
         assert env.action_space.contains(action)
-        assert np.array_equal(action, np.full(action_dim, 0.75))
+        assert np.allclose(action, expected_action)
+
         expected_mean = np.full(action_dim, 0.5)
         assert np.array_equal(agent_info['mean'], expected_mean)
         expected_log_std = np.full(action_dim, 0.5)
@@ -181,23 +190,23 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
         for action, mean, log_std in zip(actions, agent_infos['mean'],
                                          agent_infos['log_std']):
             assert env.action_space.contains(action)
-            assert np.array_equal(action, np.full(action_dim, 0.75))
+            assert np.allclose(action, expected_action)
             assert np.array_equal(mean, expected_mean)
             assert np.array_equal(log_std, expected_log_std)
 
     def test_is_pickleable(self):
         env = TfEnv(DummyBoxEnv(obs_dim=(1, ), action_dim=(1, )))
         with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy_with_model.GaussianGRUModel'),
+                         'gaussian_gru_policy.GaussianGRUModel'),
                         new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicyWithModel(
-                env_spec=env.spec, state_include_action=False)
+            policy = GaussianGRUPolicy(env_spec=env.spec,
+                                       state_include_action=False)
 
         env.reset()
         obs = env.reset()
 
-        with tf.compat.v1.variable_scope(
-                'GaussianGRUPolicyWithModel/GaussianGRUModel', reuse=True):
+        with tf.compat.v1.variable_scope('GaussianGRUPolicy/GaussianGRUModel',
+                                         reuse=True):
             return_var = tf.compat.v1.get_variable('return_var')
         # assign it to all one
         return_var.load(tf.ones_like(return_var).eval())
@@ -210,6 +219,7 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
 
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             policy_pickled = pickle.loads(p)
+            # yapf: disable
             output2 = sess.run(
                 policy_pickled.model.networks['default'].mean,
                 feed_dict={
@@ -217,3 +227,4 @@ class TestGaussianLSTMPolicyWithModel(TfGraphTestCase):
                                                  [obs.flatten()]]
                 })
             assert np.array_equal(output1, output2)
+            # yapf: enable
